+++ apps/org-sample-api/Hosting/Infrastructure/Migrations/Roles/20260208215715_SyncRolesModel20260208.cs
@@ -1,0 +1,36 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace OrgSampleApi.Hosting.Infrastructure.Migrations.Roles
+{
+    /// <inheritdoc />
+    public partial class SyncRolesModel20260208 : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<string>(
+                name: "Metadata",
+                table: "OrgSample_AuditEntries",
+                type: "text",
+                nullable: true,
+                oldClrType: typeof(string),
+                oldType: "jsonb",
+                oldNullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AlterColumn<string>(
+                name: "Metadata",
+                table: "OrgSample_AuditEntries",
+                type: "jsonb",
+                nullable: true,
+                oldClrType: typeof(string),
+                oldType: "text",
+                oldNullable: true);
+        }
+    }
+}